:truck: Moved YAMLSerialisableModel to serialisers
