+++ data_safe_haven/config/azure_serialisable_model.py
@@ -4,7 +4,7 @@ from typing import Any, ClassVar, TypeVar
 
 from data_safe_haven.context import Context
 from data_safe_haven.external import AzureApi
-from data_safe_haven.utility import YAMLSerialisableModel
+from data_safe_haven.serialisers import YAMLSerialisableModel
 
 T = TypeVar("T", bound="AzureSerialisableModel")
