+++ data_safe_haven/serialisers/__init__.py
@@ -1,0 +1,5 @@
+from .yaml_serialisable_model import YAMLSerialisableModel
+
+__all__ = [
+    "YAMLSerialisableModel",
+]
